+++ CHANGELOG.md
@@ -97,6 +97,7 @@
 
 * Integrated information about the number of asked questions and given answers by the users and also the first and last usage of the bot.
 * Updated to Project template version `4.6.2`.
+* Updated queries that uses cardinality aggregation adding the precision threshold parameter.
 
 ### 2.2.0